Improve agent services reference discoverability (#61773)
The changes in #44327 moved all reference guides in the docs to a
dedicated sidebar section as part of an effort to simplify the left
navigation sidebar. This included moving reference guides from sidebar
sections related to enrolling Teleport resources into an Agent Services
subsection of the References section.

However, the Agent Services Reference section has become difficult to
find, and the docs are now organized according to Teleport product
rather than the purpose of each guide. This change returns these
reference guides to their feature-based organization.