Remove repetition in sidebar labels (#61782)
Backports #61166, #61193, and #61252

In some sidebar sections in which multiple entries unnecessarily repeat
words, remove the repetitive words to make the sidebar easier to read.

As part of this work, add docs subsections for pages that we currently
demarcate from the rest of a subsection using repeated words. This way,
we can describe commonalities of these guides using a section index
page, and make it easier to scale out these sections as we add related
guides:

- Key management service guides
- Automatic labeling
- Managing Resources with IaC
- SSO
- Machine & Workload Identity access guides
- Top-level MWI guides
- Access Requests
- Identity Governance Okta integration
- Identity Governance integrations, including a new sidebar section for
  Entra ID integration setup guides
- Enroll Resources Joining Agents
- Kubernetes Auto-Discovery
- Protecting cloud APIs
- Application access guides, including a new sidebar section for
  enrolling different kinds of applications
- Top-level application access guides
- Enrolling databases, including new subsections for RDS Proxy and
  RDS/Aurora databases
- New subsection for enrolling MCP servers

Also make minor page title changes so all entries in a sidebar have the
same grammatical structure.