Pass `navigator.maxTouchPoints` from Web UI to auth service (#61777)
* DocumentAuthorizeWebSession: Call tshd client directly

* Add api.postWithOptions for sending POST with custom headers

There's already post and postFormData. Similar to deleteWithOptions,
instead of adding just another positional arg with custom headers to
either post or postFormData, I've decided to create a third function
with a more flexible API.

We could deprecate the other two functions but I don't know when I'll
backport this change to v18 yet. The post function should be enough for
95% of cases.

* Send Max-Touch-Points header to endpoints which lead to Device Trust prompt

* Pass max touch points to auth service

* Add iOS and iPadOS as new OSType values

* Hardcode header name

* Mention availability of `ForwardedClientMetadata.MaxTouchPoints`

* Remove new OSes from OSType for now

* Remove leftover logger

* Change fake UA in tests to be more Surface-like instead of iPad-like

* Add separate TestCreateWebSession subtest for iPadOS