Generate docs reference for the db_server resource
Add the resource to the generator config and run the generator.

A bug in the generator causes it to split `ElastiCache` despite the
string being in the list of permitted camel-case exceptions. This change
fixes by bug by checking each word in an H2 section against the
exception list before potentially splitting it.