implement app server watcher
This commit adds a new watcher for app server. This follows the
same pattern as the node watcher. This speeds up app resolve
time significantly and reduces the amount of cloning done for AppServer.

By default the max staleness time for app servers is 1 min, matching
that of node watcher.

This commit also changes the app matcher to take a readonly variant
to make it compliant with generic watcher. This also means the
checks that require context have been refactored to a seperate callsite.

Changelog: Improved application server resolution times for large number of applications.