gitlab: remove compatibility layer for client <1.2.0 (#7671)
This PR removes the compatibility layer introduced in #7670 to support
both >=v1.2.0 and <=0.159.0.
With the upgrade of master to v1.2.0, we no longer need to keep the
compatibility layer and we can drop typecasts since they are no longer
needed.

Related to #7670
Related to https://github.com/gravitational/teleport/pull/61915

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

Source-commit: gravitational/teleport.e@1f3a3979e5d373ce69e1e55cb6ea178a4b72f17b
Source-pull-requests: gravitational/teleport.e#7671