Add resource reference generator camel case config
The resourcer reference generator splits camel-cased type names in order
to create section headings. However, for terms that are already camel
cased, such a `MySQL`, the splitting logic creates headings that make
type names difficult to reason about.

This change allows us to pass a configurable list of camel-cased
keywords and abbreviations to the generator. The generator only adds
spaces after these strings, not within them. While the generator
previously preserved abbreviations automatically, we now need to pass
abbreviatons to the exception list.

This change adds camel case exceptions to the configuration in order to
preserve the two pages that we currently generate.