Remove error allocations in HeartbeatV2 (#62012)
There were four errors defined, one globally, three in
`(HeartbeatV2) run` that existed for the sole purpose of
providing an error to the callback which is used to change
the process state. At no point is the error displayed, it is
only inspected to determine whether the process should be healthy
or degraded. Each error was constructed with trace.Errorf which
results in capturing a stacktrace.

While this is may be unnoticable for a process that only has a
single heartbeat, i.e. only ssh_service enabled, this shows up
in profiles of services that orchestrate multiple heartbeats,
i.e. an app_service that matches N applications.

To reduce memory consumption all errors have been converted to
zero sized types that produce a static string. Additionally
the icsUnavailable error was removed completely because it
was never actually consumed.