Reduce allocations in app server watcher (#62014)
The slice of resources receivied from the watcher contains a copy
of all the items known by the watcher. The app server then cloned
each app via types.Application.Copy to get a handle to a types.AppV3
so that the public address could be set if it was empty. The cloned
apps are then added to a _new_ slice which is provided to the reconciler.

To reduce memory consumption, types.Application.SetPublicAddr was added
to make it possible to set the address without copying an app. The slice
from the resource watcher is now reused instead of allocating a new slice
for the amended applications.