[v18] MWI: Fix static keypair creation with `--static-key-path` (#61947)
* MWI: Fix static keypair creation with `--static-key-path`

Passing the `--static-key-path` flag was broken due to an improper
error check that was intended to allow nonexistent files, but then
incorrectly tried to extract a key from an empty byte array.

This fix skips parsing if the file is not found (which is allowed)
and allows it to simply generate a new key.

Additionally, this fixes a reference to `--static-key` in the CLI
help text, which was renamed to `--static-key-path`.

* Add test for static keypair generation

* Fix imports