Generate a reference for discovery_config (#62130)
See #61931

Edit the resource reference generator config to support the
`discovery_config` resource and generate a reference guide to the
resource's fields.

Also perform some minor cleanup of the resource reference generator
config file.

Note that, in order to add the `Metadata` field to the reference, this
change fixes a bug in which it only adds reference entries based on the
fields of a given declaration, not the fields included in any embedded
struct fields of the declaration. As part off this work, clarify
function names, renaming `handleEmbeddedStructFields` to
`allFieldsForDecl` and `getRawTypes` to `typeForDecl`.

Also remove a list of possible function declarations that the generator
used in an earlier iteration but no longer uses.