Update the Splunk audit event export guide
Closes #57002

Incorporate instructions from the GitHub discussion #56918. These allow
the guide to work with more recent versions of Splunk, since the
Universal Forwarder no longer supports the instructions in the current
guide.

This change also removs impersonation and Machine ID notes from the
"Next steps" section since we already branch on the decision to use
Machine ID earlier in the guide.