Fill in missing changelog release notes (#62215)
Add release notes that are present in the `branch/v18` and `branch/v17`
versions of the CHANGELOG but not in `master`. This will prepare us for
making the `master` CHANGELOG our record of all release notes while each
release branch's CHANGELOG contains only release notes for its
associated major version.