Speed up lib/srv/regular tests (#62337) (#62357)
The majority of the tests were spent cleaning up resources. This
was a result of calling Shutdown on the authtest.Server which
gracefully terminated connections. This was used because it is
the only method exposed to clean up, though for these tests
there is no need for graceful termination. `(authtest.Server) Close`
was added to forcefully close resources and is now used by the
test fixture created in the lib/srv/regular tests.

Several tests were also manually creating a upack, which creates
a user, a role, and generates certificates for the current user.
The same user, role, and credentials are already being generated
for the current user in `newCustomFixture` which all tests end up
calling in one way or another. Some tests were even creating a
upack for an additional user that was never used in the actual test.
All unused and additional upacks have been removed so that tests
only generate credentials as few times as possible.

The fixture has also had its dynamic labels removed and now defaults
to disabling session recordings. Any tests which require labels can
specify them directly instead of all tests spinning up a goroutine
to periodically calculate labels for no reason in tests. Similarly,
tests that actually consume recordings can opt into them instead
of wasting resources for all tests.