teleterm: Do not construct query params by hand
This messes up the encoding. When interpolating redirect_uri within a
string, the URL is simply not going to get encoded. This means that if
redirect_uri is something like "https://example.com/hello?foo=1&bar=2",
"bar=2" is going to be interpreted as one of the query param of the
returned URL, not as one of the query params of redirect_uri.