More performant resource discovery and reconciliation
Today, most reconciliation is done by creating a watcher that
is managed in a goroutine. Any time a single resource changes,
the entire new set of the resources and the old set of resources
is analyzed by a reconciler running in another goroutine. This
requires careful coordanation of the two objects and consumes
excessive memory constantly making copies.

The new ResourceMonitor achieves the same behavior within a single
object. It reduces the number of goroutines and the number of copies
of a resource set from three down to one for the same use case.