Remove legacy CRL generation
GenerateCertAuthorityCRL is not safe to use, since it isn't
signer-aware and will do the wrong thing in clusters with multiple
signers (like those using HSMs).

Remove the fallback logic that generates CRLs on demand if one is
not found in a certificate authority.

This commit only touches code that runs in the agent or in tctl,
so it is safe for v19 because v19 clients/agents require v19+
control planes.