Clean up the Management docs section (#62293)
Backports #62067

Closes #57908

Move content out of the Admin Guides and Operations sections, which are
ambiguous. There is content throughout the docs on performing admin
tasks and operations on a cluster. Each guide in these sections has a
more thematically specific home in the docs with related guides in the
same subsection.

Reorganize the following guides:

| Guide               | New Location | Reasoning                    |
|---------------------|--------------|------------------------------|
| AWS OIDC            | Cloud APIs   | Contains `integration` info  |
| Backup/restore      | Self-Hosting | Only for self-hosting        |
| CA Rotation         | Security     | About securing a cluster     |
| Datadog Integration | Diagnostics  | About monitoring a cluster   |
| GitHub integration  | Cloud APIs   | Contains `integration` info  |
| Local Users         | RBAC Intro   | Beginner-level RBAC guide    |
| Proxy Peering       | Self-Hosting | Only for self-hosting        |
| Scaling             | Reference    | Conceptual and re: deploying |
| Self-Signed Certs   | Self-Hosting | Only for self-hosting        |
| Troubleshooting     | Diagnostics  | Related to introspection     |
| Trusted Clusters    | Self-Hosting | Only for self-hosting        |
| Uninstall Teleport  | Installation | Relevant to installation     |

Also shorten sidebar labels in order to reduce noise for readers. Remove
the `trusted-cluster-address-lookup.mdx` page, since we include the same
instructions in how-to guides throughout the docs, and this is too short
to be a guide of its own.

Move the daemon guide to the top level of Management, since the Admin
section is now gone. Unlike other guides in `management/admin`, this
guide does not have another more appropriate home.