Reintroduce tsh recordings export for RemoteFX
The tsh recordings export command exports a Windows desktop session
recording to an AVI file that can be played outside Teleport. It
was built prior to our RemoteFX support, and to this day only works
on legacy PNG recordings.

This commit allows us to export RemoteFX recordings by leveraging the
Rust RDP decoder library in tsh. This will be the first time tsh links
to Rust libraries.

Closes #45388
Requires #61858