docs: update FIPS info (#62516)
This consolidates all references to BoringCrypto and FIPS-140-2 to
a single section of the docs, using more generic FIPS terminology
elsewhere. As a result, we only need to change one place when we
update our FIPS module in the future.

In addition, mention the specific versions of BoringCrypto and
associated CMVP certificate numbers.