[v18] ci: Make node buildbox runtime-architecture specific (#57204)
Add a runtime architecture suffix to the node buildbox OCI image name so
that we do not keep rebuilding it when doing an ARM64 release. ARM64
releases are built on ARM64 hosts, but the node buildbox is for AMD64 so
the ARM64 build always has to rebuild it. Adding a runtime architecture
suffix should prevent this rebuild and make the release faster.